uprava PO

+++ app/Services/PriceOfferService/PriceOfferService.php
@@ -46,8 +46,10 @@ class PriceOfferService
             $result['items'][] = $this->priceOfferItemRepository->save($priceOfferItem->toArray(), $request['id']);
         }
 
-        $itemIdList = array_column($result['items'], 'id');
-        $this->priceOfferItemRepository->deleteNotIncluded($itemIdList, $request['id']);
+        if (!empty($result['items'])) {
+            $itemIdList = array_column($result['items'], 'id');
+            $this->priceOfferItemRepository->deleteNotIncluded($itemIdList, $request['id']);
+        }
         
         $result['id'] = $request['id'];
         $result['title'] = $request['title'] ?? '';
